backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pymongo==4.6.2
python-dotenv==1.0.0
httpx==0.27.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
faiss-cpu>=1.7.4
dashscope>=1.20.0
pypdf>=4.0.0
python-docx>=1.1.0
